move pickers to context class
